+++ src/parser/bmp/error.rs
@@ -10,13 +10,14 @@ use num_enum::TryFromPrimitiveError;
 use std::error::Error;
 use std::fmt::{Display, Formatter};
 
-#[derive(Debug, PartialEq, Clone, Copy, Eq)]
+#[derive(Debug, PartialEq, Clone, Eq)]
 pub enum ParserBmpError {
     InvalidOpenBmpHeader,
     UnsupportedOpenBmpMessage,
     UnknownTlvType,
     UnknownTlvValue,
     CorruptedBmpMessage,
+    CorruptedBgpMessage(String),
     TruncatedBmpMessage,
 }
 
@@ -41,6 +42,9 @@ impl Display for ParserBmpError {
             ParserBmpError::UnknownTlvValue => {
                 write!(f, "Unknown TLV value")
             }
+            ParserBmpError::CorruptedBgpMessage(s) => {
+                write!(f, "Corrupted BGP message: {}", s)
+            }
         }
     }
 }
@@ -55,8 +59,8 @@ impl From<std::io::Error> for ParserBmpError {
 }
 
 impl From<ParserError> for ParserBmpError {
-    fn from(_: ParserError) -> Self {
-        ParserBmpError::CorruptedBmpMessage
+    fn from(e: ParserError) -> Self {
+        ParserBmpError::CorruptedBgpMessage(e.to_string())
     }
 }
 
@@ -127,6 +131,10 @@ mod tests {
             "Corrupted BMP message"
         );
         assert_eq!(
+            ParserBmpError::CorruptedBgpMessage("test".to_string()).to_string(),
+            "Corrupted BGP message: test"
+        );
+        assert_eq!(
             ParserBmpError::TruncatedBmpMessage.to_string(),
             "Truncated BMP message"
         );
@@ -148,7 +156,7 @@ mod tests {
         );
         assert_eq!(
             ParserBmpError::from(ParserError::ParseError("test".to_string())),
-            ParserBmpError::CorruptedBmpMessage
+            ParserBmpError::CorruptedBgpMessage("Error: test".to_string())
         );
         assert_eq!(
             ParserBmpError::from(TryFromPrimitiveError::<BmpMsgType>::new(0)),